+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for OpenAuthenticode
 
+## v0.2.0 - TBD
+
+* Added support for `.dll` and `.exe` files through the `PEBinary` provider
+* Output nested signatures when a file has been signed with multiple hash algorithms
+* Renamed output property `HashAlgorithmName` to just `HashAlgorithm`
+* Added the following cmdlets
+  * [Clear-OpenAuthenticodeSignature](./docs/en-US/Clear-OpenAuthenticodeSignature.md) - removes any signatures on a file
+
 ## v0.1.0 - 2023-03-17
 
 + Initial version of the `OpenAuthenticode` module